Refactor: split model.py in loader package
