+++ openfisca_web_api/loader/commons.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+DATE_FORMAT = "%Y-%m-%d"
